uudelleennimettiin convo.py mainiksi
