+++ main.py
@@ -1,135 +1,60 @@
-# coding: utf-8
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Simple Bot to reply to Telegram messages
+# This program is dedicated to the public domain under the CC0 license.
+"""
+This Bot uses the Updater class to handle the bot.
 
-from telegram import InlineKeyboardButton, InlineKeyboardMarkup
-from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, BaseFilter, CallbackQueryHandler
-import json
-import os
-import sys
+First, a few callback functions are defined. Then, those functions are passed to
+the Dispatcher and registered at their respective places.
+Then, the bot is started and runs until we press Ctrl-C on the command line.
+
+Usage:
+Example of a bot-user conversation using ConversationHandler.
+Send /start to initiate the conversation.
+Press Ctrl-C on the command line or send a signal to the process to stop the
+bot.
+"""
+
+from telegram import (ReplyKeyboardMarkup, ReplyKeyboardRemove)
+from telegram.ext import (Updater, CommandHandler, MessageHandler, Filters, RegexHandler,
+                          ConversationHandler)
+
+import logging
 from PIL import Image
+import os
+import json
 
+
+# Lue JSON-tiedosto
+def file_read(filename):
+    try:
+        with open(filename, 'r', encoding='utf-8') as file:
+            saved_data = json.load(file)
+        file.close()
+        return saved_data
+    except FileNotFoundError:
+        print("Oh dog file not found")
+        exit(1)
+
+SETTINGS = file_read("settings.json")
+ASPECT_RATIO, CANVAS_SIZE, SEND_PHOTO = range(3)
 PREFIX = 'brd_'
 JPEG_QUALITY = 100
-DEFAULT_ASPECT_RATIO = 4/5
-CANVAS_SIZE = 1.1
+data = {}
 
 
-def handlers(updater):
-    dp = updater.dispatcher
+# Enable logging
+logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+                    level=logging.INFO)
 
-    # Tässä alla oleville komennoille annetaan aina bot ja updater argumenteiksi
-    dp.add_handler(CommandHandler('start', start))
-    dp.add_handler(CommandHandler('help', help))
-    dp.add_handler(MessageHandler(Filters.document, picture))
-    dp.add_handler(CommandHandler('custom', custom, pass_args=True))
-    dp.add_handler(CallbackQueryHandler(button))
+logger = logging.getLogger(__name__)
 
 
 def start(bot, update):
     bot.send_message(chat_id=update.message.chat_id,
                      text="Hello! Send me an uncompressed picture!")
-
-
-def help(bot, update):
-    help_text = "This bot will save an uncompressed image you send it until you click on one of the buttons or use the /custom command. "\
-                "Only the latest image you send will stay saved.\n\n"\
-                "The /custom command takes three agruments. The first two make up the aspect ratio of the final image. "\
-                "The third one defines the canvas size. E.g. '/custom 4 5 1.5' first fills in the canvas (with white pixels) so that the picture "\
-                "will be in the correct aspect ratio. Then the program scales the canvas behind the image by the third argument, which will give "\
-                "the picture white frames. The image is then saved and sent to you."
-
-    bot.send_message(chat_id=update.message.chat_id, text=help_text)
-
-def check_file(fname):
-    return os.path.isfile(fname)
-
-
-def picture(bot, update):
-    img_id = update.message.document.file_id
-    filename = str(update.message.from_user.id) + '.jpeg'
-
-    imgFile = bot.get_file(img_id)
-    imgFile.download(filename)
-
-    keyboard = [[InlineKeyboardButton("1:1", callback_data=1),
-                 InlineKeyboardButton("4:5", callback_data=4/5)],
-                [InlineKeyboardButton("16:9", callback_data=16/9),
-                 InlineKeyboardButton("9:16", callback_data=9/16)],
-                [InlineKeyboardButton("9:19", callback_data=9/19)]]
-
-    reply_markup = InlineKeyboardMarkup(keyboard)
-    update.message.reply_text('Select your desired aspect ratio. Canvas size defaults to 1.1.\n'
-                              'You can also give custom values by entering:\n'
-                              '/custom A B C\n'
-                              'where A:B is aspect ratio(e.g. "3 2" = 3:2)\n'
-                              'and C is canvas size(e.g. 1.1)', reply_markup=reply_markup)
-
-
-def delete_pictures(update):
-    chat_id = update.message.chat.id
-    filename = str(chat_id) + '.jpeg'
-
-    os.remove(filename)
-    os.remove((PREFIX + filename))
-
-
-def custom(bot, update, args):
-    chat_id = update.message.chat.id
-    filename = str(chat_id) + '.jpeg'
-
-    try:
-        aspect_ratio = float(int(args[0]) / int(args[1]))
-        canvas_size = float(args[2])
-    except:
-        bot.send_message(chat_id=update.message.chat_id,
-                         text="Incorrect arguments. Try again.")
-        return
-
-    if not 0.2 < aspect_ratio < 5 or not 0 <= canvas_size <= 3:
-        bot.send_message(chat_id=update.message.chat_id,
-                         text="Too extreme values. Try again.")
-        return
-
-    if check_file(filename):
-        brd_pic = borderify(filename, aspect_ratio,
-                            canvas_size, (255, 255, 255))
-    else:
-        bot.send_message(chat_id=update.message.chat_id,
-                         text="You need to send me a picture first.")
-        return
-
-    brd_pic.save(os.path.join(PREFIX+filename), 'JPEG',
-                 quality=JPEG_QUALITY, optimize=True)
-    bot.send_document(chat_id=chat_id, document=open(
-        (PREFIX + filename), 'rb'))
-
-    delete_pictures(update)
-
-
-def button(bot, update):
-    query = update.callback_query
-    chat_id = query.message.chat.id
-    filename = str(chat_id) + '.jpeg'
-    aspect_ratio = float(query.data)
-
-    bot.edit_message_text(text="This will take just a second...",
-                          chat_id=query.message.chat_id,
-                          message_id=query.message.message_id)
-
-    if check_file(filename):
-        brd_pic = borderify(filename, aspect_ratio,
-                            CANVAS_SIZE, (255, 255, 255))
-    else:
-        bot.edit_message_text(text="You need to send me a picture first.",
-                              chat_id=query.message.chat_id,
-                              message_id=query.message.message_id)
-        return
-
-    brd_pic.save(os.path.join(PREFIX+filename), 'JPEG',
-                 quality=JPEG_QUALITY, optimize=True)
-    bot.send_document(chat_id=chat_id, document=open(
-        (PREFIX + filename), 'rb'))
-
-    delete_pictures(update)
 
 
 def borderify(name, aspect_ratio=4/5, margin_ratio=1.1, background_color=(255, 255, 255)):
@@ -160,25 +85,119 @@ def borderify(name, aspect_ratio=4/5, margin_ratio=1.1, background_color=(255, 2
     return bg
 
 
-# Lue JSON-tiedosto
-def file_read(filename):
-    try:
-        with open(filename, 'r', encoding='utf-8') as file:
-            saved_data = json.load(file)
-        file.close()
-        return saved_data
-    except FileNotFoundError:
-        print("Oh dog file not found")
-        exit(1)
+def photo(bot, update):
+    user = update.message.from_user
+    user_id = update.message.from_user.id
+    logger.info("%s sent a photo", user.username)
+
+    photo_file = bot.get_file(update.message.document.file_id)
+    photo_file.download(str(user_id) + '.jpeg')
+
+    reply_keyboard = [['1/1', '4/5', '16/9'],['9/16', '9/19', '2/1']]
+    update.message.reply_text('What aspect ratio are you looking for?', reply_markup=ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True))
+
+    return ASPECT_RATIO
+
+
+def aspect_ratio(bot, update):
+    user = update.message.from_user
+    user_id = update.message.from_user.id
+    global data
+
+    data[user_id] = {"ar": update.message.text}
+
+    logger.info("Aspect ratio for %s: %s", user.username, update.message.text)
+
+    reply_keyboard = [['1', '1.02', '1.05'],['1.1', '1.2', '1.3']]
+    update.message.reply_text('How large do you want your canvas to be?', reply_markup=ReplyKeyboardMarkup(reply_keyboard, one_time_keyboard=True))
+
+    return CANVAS_SIZE
+
+
+def canvas_size(bot, update):
+    user = update.message.from_user
+    user_id = update.message.from_user.id
+    chat_id = update.message.chat.id
+    logger.info("Canvas size for %s: %s", user.username, update.message.text)
+
+    global data
+
+    data[user_id]["cs"] = update.message.text
+
+    ar = eval(data[user_id]["ar"])
+    cs = float(data[user_id]["cs"])
+
+    filename = str(user_id) + '.jpeg'
+
+    pic = borderify(filename, ar, cs, (255, 255, 255))
+
+    pic.save(os.path.join(PREFIX+filename), 'JPEG', quality=JPEG_QUALITY, optimize=True)
+    bot.send_document(chat_id=chat_id, document=open((PREFIX + filename), 'rb'))
+
+    delete_pictures(update)
+
+    return ConversationHandler.END
+
+
+def cancel(bot, update):
+    user = update.message.from_user
+    user_id = update.message.from_user.id
+
+    logger.info("User %s canceled the conversation.", user.username)
+    update.message.reply_text('Bye! I hope we can talk again some day.',
+                              reply_markup=ReplyKeyboardRemove())
+
+    return ConversationHandler.END
+
+
+def error(bot, update, error):
+    """Log Errors caused by Updates."""
+    logger.warning('Update "%s" caused error "%s"', update, error)
+
+def delete_pictures(update):
+    chat_id = update.message.chat.id
+    filename = str(chat_id) + '.jpeg'
+
+    os.remove(filename)
+    os.remove((PREFIX + filename))
 
 
 def main():
-    updater = Updater(token=SETTINGS["tg_token"])
-    handlers(updater)
+    # Create the EventHandler and pass it your bot's token.
+    updater = Updater(SETTINGS["tg_token"])
 
+    # Get the dispatcher to register handlers
+    dp = updater.dispatcher
+
+    # Add conversation handler with the states ASPECT_RATIO, CANVAS_SIZE, LOCATION and BIO
+    conv_handler = ConversationHandler(
+        entry_points=[MessageHandler(Filters.document, photo)],
+
+        states={
+            # TODO custom values
+            ASPECT_RATIO: [RegexHandler('^(1/1|4/5|16/9|9/16|9/19|2/1)$', aspect_ratio)],
+
+            CANVAS_SIZE: [RegexHandler('^(1|1.02|1.05|1.1|1.2|1.3)$', canvas_size)]
+        },
+
+        fallbacks=[CommandHandler('cancel', cancel)]
+    )
+
+    dp.add_handler(conv_handler)
+    dp.add_handler(CommandHandler('start', start))
+
+
+    # log all errors
+    dp.add_error_handler(error)
+
+    # Start the Bot
     updater.start_polling()
 
+    # Run the bot until you press Ctrl-C or the process receives SIGINT,
+    # SIGTERM or SIGABRT. This should be used most of the time, since
+    # start_polling() is non-blocking and will stop the bot gracefully.
+    updater.idle()
 
-SETTINGS = file_read("settings.json")
 
-main()
+if __name__ == '__main__':
+    main()
